Built components for the different stages of the Tea Round Picker including: -Add Drink -Create User -More Users Required
Built a component that selects these stages based on the state of the site called CurrentStage

The App component now embeds the CurrentStage component and simply shows a title above it.

Cleaned up App.css to get rid of unused styles and classes from the template and added an inputBox class for the input boxes.